requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
python-dotenv>=1.0.0
Jinja2>=3.1.3
